Move modal into separate module
